finger: Decrease timeout length during connect()
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=572211

In cases where a name server is answering with A as well as AAAA records,
but the system to be queried has lost a corresponding address, the TCP
handshake timeout will cause a long delay before allowing the query of
the next address family, or the next address in general.

The use of a trivial signal handler for SIGALRM allows the reduction
of this timeout, thus producing better responsiveness for the interactive
user of the Finger service.

Imported from Debian 06-572211-decrease-timeout.patch