finger: Fix segfault on empty gecos
If gecos is empty (or consists entirely of delemiters, as set by some
user-creation scripts), then userinfo() fails to set pn->realname,
leading to an uninitialized pointer dereference when printed. Instead,
we set it to null if there are no gecos fields.

Additionally, we update lprint() to no longer assume pn->realname is
non-NULL. That check is already made by sflag_print().